+++ src/on-posix/hw_specific.cpp
@@ -27,7 +27,7 @@ namespace RODOS {
 */
 
 extern void sigtermHandler(int sig);
-extern long myNodeNr;
+extern int32_t myNodeNr;
 
 void hwInit() {
     taskRunning = 0;
